Split `git` module into multiple ones
